Refactor ccLibAlgorithms from MainWindow into their own namespace
